moving CM_module to models_extra and deleting other_models
