+++ pareto/utilities/cm_utils/opt_utils.py
@@ -14,7 +14,7 @@
 Functions used in run_infrastructure_analysis.py
 """
 
-from pareto.other_models.CM_module.models.qcp_br import build_qcp_br
+from pareto.models_extra.CM_module.models.qcp_br import build_qcp_br
 import pyomo.environ as pyo
 from pareto.utilities.cm_utils.gen_utils import obj_fix
