+++ pareto/utilities/cm_utils/run_utils.py
@@ -18,7 +18,7 @@ This also contains the solving function.
 
 import numpy as np
 import pyomo.environ as pyo
-from pareto.other_models.CM_module.models.qcp_br import build_qcp_br
+from pareto.models_extra.CM_module.models.qcp_br import build_qcp_br
 
 from pareto.utilities.get_data import get_data
 from pareto.utilities.cm_utils.gen_utils import (
@@ -32,7 +32,7 @@ import matplotlib.pyplot as plt
 import matplotlib.patches as mpatches
 import os
 import sys
-from pareto.other_models.CM_module.operational.set_param_list import (
+from pareto.models_extra.CM_module.set_param_list import (
     set_list,
     parameter_list,
 )
